requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
certifi==2021.10.8
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.3
django-bootstrap5==21.3
djangorestframework==3.13.1
filelock==3.6.0
gunicorn==20.1.0
numpy==1.22.3
Pillow==9.0.1
pipenv==2022.1.8
platformdirs==2.5.1
psycopg2==2.9.3
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
virtualenv==20.13.3
virtualenv-clone==0.5.7
whitenoise==6.0.0
